Shop

+++ Section5/Shop_5_5/Product.cs
@@ -2,13 +2,13 @@
 {
     internal class Product
     {
-        public Product(string name, float price)
+        public Product(string name, int price)
         {
             Name = name;
             Price = price;
         }
 
         public string Name { get; private set; }
-        public float Price { get; private set; }
+        public int Price { get; private set; }
     }
 }
